Update growl display
I've changed the growl display to show all current headlines. Getting
the first top headline wasn't showing the news I had hoped. This gives
me all the current headlines, which provides a little more information.

This may prove to be too much. Maybe allow for a parameter to be passed
for number? I definitely also want to pass what sort of news we want to
receive as well.